+++ src/Models/Models.jl
@@ -18,6 +18,8 @@ using Oceananigans.AbstractOperations: AbstractOperation
 import Oceananigans: initialize!
 import Oceananigans.Architectures: architecture
 
+using Oceananigans.OutputReaders: update_field_time_series!, extract_field_timeseries
+
 architecture(model::AbstractModel) = model.architecture
 initialize!(model::AbstractModel) = nothing
 
@@ -25,11 +27,6 @@ total_velocities() = nothing
 possible_field_time_series(model) = ()
 
 import Oceananigans.TimeSteppers: reset!
-
-# This function should eventually be defined solely in OutputReaders.
-# We define it here as a place holder because Models is imported into
-# Oceananigans before OutputReaders right now.
-update_field_time_series!(::Nothing, time) = nothing
 
 # Update _all_ `FieldTimeSeries`es in an `AbstractModel`. 
 # Loop over all propery names and extract any of them which is a `FieldTimeSeries`.
@@ -49,29 +46,6 @@ function update_model_field_time_series!(model::AbstractModel, clock::Clock)
 
     return nothing
 end
-
-# Recursion for all properties 
-function extract_field_timeseries(t) 
-    prop = propertynames(t)
-    if isempty(prop)
-        return nothing
-    end
-
-    return Tuple(extract_field_timeseries(getproperty(t, p)) for p in prop)
-end
-
-# For types that do not contain `FieldTimeSeries`, halt the recursion
-NonFTS = [:Number, :AbstractArray]
-
-for NonFTSType in NonFTS
-    @eval extract_field_timeseries(::$NonFTSType) = nothing
-end
-
-# Special recursion rules for `Tuple` and `Field` types
-extract_field_timeseries(t::AbstractField)     = Tuple(extract_field_timeseries(getproperty(t, p)) for p in propertynames(t))
-extract_field_timeseries(t::AbstractOperation) = Tuple(extract_field_timeseries(getproperty(t, p)) for p in propertynames(t))
-extract_field_timeseries(t::Tuple)             = Tuple(extract_field_timeseries(n) for n in t)
-extract_field_timeseries(t::NamedTuple)        = Tuple(extract_field_timeseries(n) for n in t)
 
 function reset!(model::AbstractModel)
