+++ src/Oceananigans.jl
@@ -203,6 +203,7 @@ include("Fields/Fields.jl")
 include("AbstractOperations/AbstractOperations.jl")
 include("Advection/Advection.jl")
 include("Solvers/Solvers.jl")
+include("OutputReaders/OutputReaders.jl")
 
 # TODO: move here
 #include("ImmersedBoundaries/ImmersedBoundaries.jl")
@@ -227,7 +228,6 @@ include("Models/Models.jl")
 # Output and Physics, time-stepping, and models
 include("Diagnostics/Diagnostics.jl")
 include("OutputWriters/OutputWriters.jl")
-include("OutputReaders/OutputReaders.jl")
 include("Simulations/Simulations.jl")
 
 # Abstractions for distributed and multi-region models
@@ -249,6 +249,7 @@ using .BuoyancyModels
 using .StokesDrift
 using .TurbulenceClosures
 using .Solvers
+using .OutputReaders
 using .Forcings
 using .ImmersedBoundaries
 using .DistributedComputations
@@ -256,7 +257,6 @@ using .Models
 using .TimeSteppers
 using .Diagnostics
 using .OutputWriters
-using .OutputReaders
 using .Simulations
 using .AbstractOperations
 using .MultiRegion
