+++ src/OutputReaders/field_time_series.jl
@@ -15,7 +15,6 @@ using Oceananigans.Fields: show_location, interior_view_indices, data_summary, r
 
 using Oceananigans.Fields: boundary_conditions 
 
-using Oceananigans.OutputWriters: serializeproperty!, iteration_exists
 import Oceananigans.Fields: Field, set!, interior, indices
 import Oceananigans.Architectures: architecture
